back-endd/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
SQLAlchemy==2.0.43
alembic==1.16.5
bcrypt==4.3.0
passlib==1.7.4
python-dotenv==1.1.1
python-jose==3.5.0
requests==2.32.5